Split Rom class
