+++ msx2cas-core/src/main/java/br/com/dod/vcas/wav/Bas.java
@@ -14,7 +14,7 @@ public class Bas extends Wav {
 
     @Override
     protected void validate() throws FlowException {
-        if (this.fileLength < MIN_ENC_INPUTFILE_LENGTH || this.fileLength > MAX_ENC_INPUTFILE_LENGTH) throw FlowException.error("file_size_invalid");
+        if (this.fileLength < MIN_ENC_INPUTFILE_LENGTH) throw FlowException.error("file_size_invalid");
     }
 
     @Override
