+++ msx2cas-core/src/main/java/br/com/dod/vcas/wav/Wav.java
@@ -12,8 +12,6 @@ import br.com.dod.vcas.exception.FlowException;
 public abstract class Wav {
 
     static final long MIN_ENC_INPUTFILE_LENGTH = 5L;
-    static final long MAX_ENC_INPUTFILE_LENGTH = 32768L;
-
     static final int CAS_FILENAME_LENGTH = FileCommons.CAS_FILENAME_LENGTH;
 
     private static final char START_BIT = 0;
